requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
astropy==7.2.0
astropy-iers-data==0.2026.1.26.0.43.56
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.10.1
fastar==0.8.0
greenlet==3.3.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.45.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numba==0.62.1
numpy==2.3.5
packaging==26.0
psycopg2==2.9.11
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyerfa==2.0.1.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.21
pyuvdata==3.2.5
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
scipy==1.17.0
sentry-sdk==2.49.0
setuptools==80.10.2
setuptools-scm==9.2.2
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.45
sqlmodel==0.0.31
starlette==0.50.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvtools==0.2.0
watchfiles==1.1.1
websockets==16.0
asyncpg==0.31.0